Move ssl and pooling into a separate source files
