+++ libs/exo-sql/src/sql/connect/mod.rs
@@ -1,2 +1,4 @@
 pub mod database_client;
 pub mod database_client_manager;
+pub mod database_pool;
+pub mod ssl_config;
